LayoutTests/platform/gtk/fast/canvas/webgl/drawElements-empty-vertex-data-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_OPERATION: drawElements: attempt to access out of bounds arrays
PASS: Unable to draw with invalid vertexAttribArray0
